Updated UI with extra functionality and better integration with Spout
Changed some of the wording and variable management in UI (now is UserInterface.py)

Added cookies-like functionality to the UI such that you have the option to save the information with Save Session checkbox. After reopening with Save Session checked off, all the information will still be there from the previous session. If you uncheck Save Session and close the UI, the session information will be deleted and upon reopening, it will be an empty UI as normal.

Integrated the UI object as an optional parameter such that popup boxes can be introduced for prompts. Only prompt which has been integrated is whether the user would like to overwrite an existing manifest file.