Added the ability to detect partial cutouts
Added the ability for Zooniverse_Datasets to detect partial cutouts and remove them automatically. The removed targets are sent to an "ignored-targets.csv" file.

Additionally, the structure of Zooniverse_Dataset was adjusted so that any child class of Zooniverse_Dataset will use its own "generateDataAndMetadataLists" function, so that specific project functionality can be implemented on a per-project basis in the future.