Added the ability to set values in CN_Dataset for the number of subdirectories and how many images per subdirectory
Sub directory limit controls the number of sub directories. For example, it is current set to 9999, so sub directories will have the form 0000, 0001, ...
Sub directory threshold is the number of images per sub directory. This doesn't mean it will exactly have that many images but rather once it goes over that amount, the next iteration will go to a new sub directory.