Added some minor changes
In order to to figure out "The Mystery of Target 15993", I had a lot of temporary bug-testing functionality for testing the validity of subjects. Some of which seem useful enough to keep for future uses, but most of the stuff was not well written for general-use cases and have been removed.